python_code/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
pydantic==2.9.0
pydantic_core==2.23.2
scikit-learn==1.5.1
scipy==1.14.0
sniffio==1.3.1
starlette==0.38.5
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.3.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
python-dotenv==1.0.1